requirements.txt
openai
transformers
requests
beautifulsoup4
streamlit
python-dotenv
